Add debug mode
Debug mode can be used to prevent the user from accessing a feature that
is only meant to be used while debugging a server written with Toldi.js.

Debug mode can be entered using the setDebugMode function, or by setting
the TOLDI_DEBUG environment variable to any value.

The current state of debug mode (enabled or disabled) can be accessed
using the getDebugMode function.